Isolated menu with class MyDrawer
