+++ ask_away/lib/main.dart
@@ -1,6 +1,6 @@
+import 'package:ask_away/Menu.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/painting.dart';
 
 import 'package:ask_away/Question.dart';
 
@@ -58,22 +58,6 @@ void createQuestion() {
   return;
 }
 
-void _userProfile() {
-  print("Profile\n");
-  return;
-}
-
-Widget _userIcon() {
-  return InkWell(
-    onTap: _userProfile,
-    child: CircleAvatar(
-      radius: 30,
-      backgroundImage: AssetImage('assets/avatar.jpg'),
-      backgroundColor: Colors.transparent,
-    ),
-  );
-}
-
 class MyHomePageState extends StatefulWidget {
   @override
   _MyHomePageState createState() => new _MyHomePageState();
@@ -99,53 +83,7 @@ class _MyHomePageState extends State<MyHomePage> {
           },
           child: qList
       ),
-      drawer: Drawer(
-        child: ListView(
-          children: <Widget>[
-            Container(
-              height: 100,
-              child: DrawerHeader(
-                child: Row(children: <Widget>[
-                  Column(children: <Widget>[
-                    _userIcon(),
-                  ]),
-                  Container(
-                    padding: new EdgeInsets.only(left: 20),
-                    child: Align(
-                        alignment: Alignment.centerLeft,
-                        child: Text('Mr. Padoru',
-                            style: TextStyle(
-                                fontWeight: FontWeight.bold, fontSize: 20))),
-                  ),
-                ]),
-              ),
-              decoration: BoxDecoration(
-                color: Colors.blue,
-              ),
-            ),
-            ListTile(
-              title: Text('Talks',
-                  style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20)),
-              onTap: () {
-                // Update the state of the app
-                // ...
-                // Then close the drawer
-                Navigator.pop(context);
-              },
-            ),
-            ListTile(
-              title: Text('Settings',
-                  style: TextStyle(fontWeight: FontWeight.bold, fontSize: 20)),
-              onTap: () {
-                // Update the state of the app
-                // ...
-                // Then close the drawer
-                Navigator.pop(context);
-              },
-            ),
-          ],
-        ),
-      ),
+      drawer: MyDrawer()
     );
   }
 }
